moved tests to the correct package
