+++ bundles/core/org.eclipse.smarthome.core.test/src/test/java/org/eclipse/smarthome/core/scheduler/DateExpressionTest.java
@@ -1,23 +1,20 @@
-package org.eclipse.smarthome.core.common;
+package org.eclipse.smarthome.core.scheduler;
 
 import static org.junit.Assert.assertEquals;
 
 import java.text.ParseException;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.TimeZone;
 
+import org.eclipse.smarthome.core.scheduler.DateExpression;
 import org.junit.Test;
 
-public class CronExpressionTest {
+public class DateExpressionTest {
 
     @Test(expected = ParseException.class)
     public void garbageString() throws ParseException {
-        new CronExpression("blahblahblah");
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void dayOfWeekAndMonth() throws ParseException {
-        new CronExpression("* * * 1 * 1");
+        new DateExpression("blahblahblah");
     }
 
     @Test
@@ -25,14 +22,19 @@ public class CronExpressionTest {
 
         Calendar cal = Calendar.getInstance();
         cal.set(2016, 0, 1, 0, 0, 0); // set to Jan 1st 2016, 00:00
+        cal.setTimeZone(TimeZone.getTimeZone("GMT+00"));
         Date startDate = cal.getTime();
 
-        // Fire at 10:15am on the third Friday of every month
-        CronExpression expr = new CronExpression("0 15 10 ? * 6#3", startDate);
+        DateExpression expr = new DateExpression("2016-01-31T00:00:00+00:00");
+        expr.setStartDate(startDate);
+        expr.setTimeZone(TimeZone.getTimeZone("GMT+00"));
 
         Date nextDate = expr.getTimeAfter(startDate);
 
-        cal.set(2016, 0, 15, 10, 15, 0);
+        cal.set(2016, 0, 31, 0, 0, 0);
+        cal.set(Calendar.MILLISECOND, 0);
+
+        cal.setTimeZone(TimeZone.getTimeZone("GMT+00"));
         Date checkDate = cal.getTime();
 
         assertEquals(checkDate, nextDate);
@@ -43,14 +45,18 @@ public class CronExpressionTest {
 
         Calendar cal = Calendar.getInstance();
         cal.set(2016, 0, 1, 0, 0, 0); // set to Jan 1st 2016, 00:00
+        cal.setTimeZone(TimeZone.getTimeZone("GMT+00"));
         Date startDate = cal.getTime();
 
-        // Fire at 10:15am on every last friday of every month during the years 2016 to 2020
-        CronExpression expr = new CronExpression("0 15 10 ? * 6L 2016-2020", startDate);
+        DateExpression expr = new DateExpression("2016-01-31T00:00:00+00:00");
+        expr.setStartDate(startDate);
+        expr.setTimeZone(TimeZone.getTimeZone("GMT+00"));
 
         Date nextDate = expr.getFinalFireTime();
 
-        cal.set(2020, 11, 25, 10, 15, 0);
+        cal.set(2016, 0, 31, 0, 0, 0);
+        cal.set(Calendar.MILLISECOND, 0);
+        cal.setTimeZone(TimeZone.getTimeZone("GMT+00"));
         Date checkDate = cal.getTime();
 
         assertEquals(checkDate, nextDate);
